refactor: Centralize mutualized prop type in a file
Avoid can't access lexical declaration 'X' before initialization
