+++ src/components/Form/FieldInputTextarea.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import TextField from 'cozy-ui/transpiled/react/TextField'
 
-import { FieldInputWrapperPropTypes } from './FieldInputWrapper'
+import { FieldInputWrapperPropTypes } from './fieldInputType'
 
 const FieldInputTextarea = props => {
   const { isMultiline, ...rest } = props
