+++ src/components/Form/FieldInputWrapper.jsx
@@ -1,8 +1,7 @@
-import PropTypes from 'prop-types'
 import React from 'react'
 
+import { FieldInputWrapperPropTypes } from './fieldInputType'
 import { getFieldInput } from '../../helpers/getFieldInput'
-import { fieldInputAttributes } from '../ContactCard/ContactFields/ContactFieldsProptypes'
 
 const FieldInputWrapper = ({ input, attributes, ...props }) => {
   const { labelProps, ...restOfAttributes } = attributes || {}
@@ -17,27 +16,6 @@ const FieldInputWrapper = ({ input, attributes, ...props }) => {
       InputLabelProps={labelProps}
     />
   )
-}
-
-export const FieldInputWrapperPropTypes = {
-  // See official documentation for more information: https://final-form.org/docs/react-final-form/types/FieldRenderProps
-  input: PropTypes.object,
-  //
-  attributes: fieldInputAttributes,
-  // Destructuring props
-  // See official documentation for more information: https://final-form.org/docs/react-final-form/types/FieldRenderProps
-  meta: PropTypes.object,
-  //
-  id: PropTypes.string,
-  label: PropTypes.string,
-  required: PropTypes.bool,
-  variant: PropTypes.string,
-  customLabelOptions: PropTypes.shape({
-    hide: PropTypes.bool,
-    defaultType: PropTypes.string.isRequired,
-    defaultLabel: PropTypes.string.isRequired
-  }),
-  fullWidth: PropTypes.bool
 }
 
 FieldInputWrapper.propTypes = FieldInputWrapperPropTypes
